# app/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FAIRS Roulette Backend",
    "description": "FastAPI backend",
    "version": "2.3.0"
  },
  "paths": {
...
    "/api/training/start": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Start Training",
        "operationId": "start_training_api_training_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/resume": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Resume Training",
        "operationId": "resume_training_api_training_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/training/checkpoints/{checkpoint}/metadata": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Checkpoint Metadata",
        "operationId": "get_checkpoint_metadata_api_training_checkpoints__checkpoint__metadata_get",
        "parameters": [
          {
            "name": "checkpoint",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkpoint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingCheckpointMetadataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/checkpoints/{checkpoint}": {
      "delete": {
        "tags": [
          "training"
        ],
        "summary": "Delete Checkpoint",
        "operationId": "delete_checkpoint_api_training_checkpoints__checkpoint__delete",
        "parameters": [
          {
            "name": "checkpoint",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkpoint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingCheckpointDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/training/jobs/{job_id}": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Training Job Status",
        "operationId": "get_training_job_status_api_training_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "training"
        ],
        "summary": "Cancel Training Job",
        "operationId": "cancel_training_job_api_training_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobCancelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobCancelResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "success",
          "message"
        ],
        "title": "JobCancelResponse"
      },
      "JobStartResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "poll_interval": {
            "type": "number",
            "title": "Poll Interval",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "job_type",
          "status",
          "message"
        ],
        "title": "JobStartResponse"
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "job_type": {
            "type": "string",
            "title": "Job Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "poll_interval": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poll Interval"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "job_type",
          "status",
          "progress"
        ],
        "title": "JobStatusResponse"
      },
...
      "ResumeConfig": {
        "properties": {
          "checkpoint": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Checkpoint"
          },
          "additional_episodes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Additional Episodes",
            "default": 10
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "checkpoint"
        ],
        "title": "ResumeConfig",
        "description": "Configuration for resuming a training session from a checkpoint."
      },
      "TrainingCheckpointDeleteResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "TrainingCheckpointDeleteResponse"
      },
...
      "TrainingCheckpointMetadataResponse": {
        "properties": {
          "checkpoint": {
            "type": "string",
            "title": "Checkpoint"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "checkpoint",
          "summary"
        ],
        "title": "TrainingCheckpointMetadataResponse"
      },
      "TrainingConfig": {
        "properties": {
          "perceptive_field_size": {
            "type": "integer",
            "maximum": 1024.0,
            "minimum": 1.0,
            "title": "Perceptive Field Size",
            "default": 64
          },
          "qnet_neurons": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Qnet Neurons",
            "default": 64
          },
          "embedding_dimensions": {
            "type": "integer",
            "minimum": 8.0,
            "title": "Embedding Dimensions",
            "default": 200
          },
          "exploration_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Exploration Rate",
            "default": 0.75
          },
          "exploration_rate_decay": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Exploration Rate Decay",
            "default": 0.995
          },
          "minimum_exploration_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Minimum Exploration Rate",
            "default": 0.1
          },
          "discount_rate": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Discount Rate",
            "default": 0.5
          },
          "model_update_frequency": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Model Update Frequency",
            "default": 10
          },
          "bet_amount": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Bet Amount",
            "default": 10
          },
          "initial_capital": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Initial Capital",
            "default": 1000
          },
          "dynamic_betting_enabled": {
            "type": "boolean",
            "title": "Dynamic Betting Enabled",
            "default": false
          },
          "bet_strategy_model_enabled": {
            "type": "boolean",
            "title": "Bet Strategy Model Enabled",
            "default": false
          },
          "bet_strategy_fixed_id": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Bet Strategy Fixed Id",
            "default": 0
          },
          "strategy_hold_steps": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Strategy Hold Steps",
            "default": 1
          },
          "bet_unit": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bet Unit"
          },
          "bet_max": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Bet Max"
          },
          "bet_enforce_capital": {
            "type": "boolean",
            "title": "Bet Enforce Capital",
            "default": true
          },
          "dataset_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Id"
          },
          "use_data_generator": {
            "type": "boolean",
            "title": "Use Data Generator",
            "default": false
          },
          "num_generated_samples": {
            "type": "integer",
            "minimum": 100.0,
            "title": "Num Generated Samples",
            "default": 10000
          },
          "sample_size": {
            "type": "number",
            "maximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Sample Size",
            "default": 1.0
          },
          "validation_size": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Validation Size",
            "default": 0.2
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          },
          "episodes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Episodes",
            "default": 10
          },
          "max_steps_episode": {
            "type": "integer",
            "minimum": 100.0,
            "title": "Max Steps Episode",
            "default": 2000
          },
          "batch_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Batch Size",
            "default": 32
          },
          "learning_rate": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Learning Rate",
            "default": 0.0001
          },
          "max_memory_size": {
            "type": "integer",
            "minimum": 100.0,
            "title": "Max Memory Size",
            "default": 10000
          },
          "replay_buffer_size": {
            "type": "integer",
            "minimum": 100.0,
            "title": "Replay Buffer Size",
            "default": 1000
          },
          "training_seed": {
            "type": "integer",
            "title": "Training Seed",
            "default": 42
          },
          "checkpoint_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkpoint Name"
          },
          "use_device_gpu": {
            "type": "boolean",
            "title": "Use Device Gpu",
            "default": false
          },
          "device_id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Device Id",
            "default": 0
          },
          "use_mixed_precision": {
            "type": "boolean",
            "title": "Use Mixed Precision",
            "default": false
          },
          "jit_compile": {
            "type": "boolean",
            "title": "Jit Compile",
            "default": false
          },
          "jit_backend": {
            "type": "string",
            "minLength": 1,
            "title": "Jit Backend",
            "default": "inductor"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TrainingConfig",
        "description": "Configuration for starting a new training session."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}